Allow to mock multiple protocol in a single object
An Objective-C object can naturally implements an arbitrary number of
protocol. I ended up in a case where I had to modify my code, because
an object was expected to implement two protocols and OCMock does not
allow it. I’d love to see this code merged so that it can.

I also edited the code for nice mocks in order for the API to be
consistent. I don’t actually need it.

It should be noted that the description is not changed in the case
where there is a single protocol mocked.